+++ sequelize-utils/lib/sequelize-utils.module.ts
@@ -1,11 +1,14 @@
 import { Global, Module, OnModuleInit } from "@nestjs/common";
 import { Sequelize } from "sequelize-typescript";
-import { AfterCreateFormatDateOnlyHook } from "./hooks/after-create-format-date-only.hook";
+import {
+    AfterBulkCreateFormatDateOnlyHook,
+    AfterCreateFormatDateOnlyHook
+} from "./hooks/after-create-format-date-only.hook";
 import { AfterFindFormatDateOnlyHook } from "./hooks/after-find-format-date-only.hook";
 import { SequelizeTransactionService } from "./transactions/sequelize-transaction.service";
 
+/** @deprecated: Use SequelizeUtilsModule instead */
 @Module({})
-/** @deprecated Use SequelizeUtilsModule instead */
 export class SequelizeUtilsHooksModule implements OnModuleInit {
     constructor(private sequelize: Sequelize) {}
 
@@ -24,7 +27,9 @@ export class SequelizeUtilsModule implements OnModuleInit {
     constructor(private sequelize: Sequelize) {}
 
     public onModuleInit(): void {
+        this.sequelize.addHook("afterBulkCreate", "transform-date-only", AfterBulkCreateFormatDateOnlyHook);
         this.sequelize.addHook("afterCreate", "transform-date-only", AfterCreateFormatDateOnlyHook);
         this.sequelize.addHook("afterFind", "transform-date-only", AfterFindFormatDateOnlyHook);
+        this.sequelize.addHook("afterUpdate", "transform-date-only", AfterFindFormatDateOnlyHook);
     }
 }